+++ pennylane/vqe/vqe.py
@@ -474,6 +474,9 @@ class ExpvalCost:
         optimize=False,
         **kwargs,
     ):
+        if kwargs.get("measure", "expval") != "expval":
+            raise ValueError("ExpvalCost can only be used to construct sums of expectation values.")
+
         coeffs, observables = hamiltonian.terms
 
         self.hamiltonian = hamiltonian